Refactor collections manage page into dedicated hook
